+++ NEWS.md
@@ -5,6 +5,8 @@
   * New function `read_airports()` to read data on all public and private airports. [Closes #4](https://github.com/ipeaGIT/flightsbr/issues/4) and [Closes #9](https://github.com/ipeaGIT/flightsbr/issues/9).
   * Function `read_flighs()` now takes `date` input in the format `yyyymm` or `yyyy`. The the date is a 4-digit number, the function now downloads the data of all months in that year. [Closes #1](https://github.com/ipeaGIT/flightsbr/issues/1).
   * Function `read_flighs()` now automatically detects and checks the latest flights data available. [Closes #13](https://github.com/ipeaGIT/flightsbr/issues/13).
+  * Function `latest_flights_date()` to check the date of the latest flight data available. [Closes #16](https://github.com/ipeaGIT/flightsbr/issues/16).
+  new function to check the date of the latest flight data available
   * new internal support functions:
     * `split_date()`: Split a date from yyyymmm to year yyyy and month mm
     * `check_date()`: Check whether date input is acceptable